Add task creator component and refactor tasks
