+++ tasks/src/components/atoms/TaskDescriptionInput/index.js
@@ -1,11 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-const TaskDescriptionInput = ({ value, handleChange, handleSubmit }) => {
+const TaskDescriptionInput = ({ value, handleChange }) => {
   return (
     <input
       value={value}
-      onBlur={handleSubmit}
 			onChange={handleChange}
     />
   )
@@ -13,8 +12,7 @@ const TaskDescriptionInput = ({ value, handleChange, handleSubmit }) => {
 
 TaskDescriptionInput.propTypes = {
   value: PropTypes.string,
-  handleChange: PropTypes.func.isRequired,
-  handleSubmit: PropTypes.func.isRequired
+  handleChange: PropTypes.func.isRequired
 }
 
 export default TaskDescriptionInput
