+++ tasks/src/components/organisms/TaskCreator/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "index.js"
+}
