+++ tasks/src/components/organisms/TaskList/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "index.js"
+}
