+++ tasks/src/components/organisms/Tasks/index.js
@@ -1,59 +1,15 @@
 import React, { Component } from 'react'
-import { connect } from 'react-redux'
-import PropTypes from 'prop-types'
 
-import {
-  fetchTasks, updateTask, deleteTask
-} from '../../../ducks/modules/tasks'
+import TaskList from '../TaskList/index'
+import TaskCreator from '../TaskCreator/index'
 
-import TaskItems from '../../molecules/TaskItems/index'
-
-export class Tasks extends Component {
-  componentDidMount() {
-    this.props.fetchTasks()
-  }
-
-  updateTask = (payload) => {
-    this.props.updateTask(payload)
-  }
-
-  deleteTask = (payload) => {
-    this.props.deleteTask(payload)
-  }
-
+export default class Tasks extends Component {
   render() {
     return (
-      <TaskItems
-        items={this.props.items}
-        updateTask={this.updateTask}
-        deleteTask={this.deleteTask}
-      />
+      <div>
+        <TaskList />
+        <TaskCreator />
+      </div>
     )
   }
 }
-
-Tasks.propTypes = {
-  items: PropTypes.array.isRequired,
-  fetchTasks: PropTypes.func.isRequired,
-  updateTask: PropTypes.func.isRequired,
-  deleteTask: PropTypes.func.isRequired
-}
-
-const mapStateToProps = (state) => {
-  return {
-    items: state.tasks.items
-  }
-}
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    fetchTasks: () => dispatch(fetchTasks()),
-    updateTask: (payload) => dispatch(updateTask(payload)),
-    deleteTask: (payload) => dispatch(deleteTask(payload))
-  }
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(Tasks)
